+++ scifit/__init__.py
@@ -1,1 +1,6 @@
+import logging
+
 __version__ = "0.1.10"
+
+
+logger = logging.getLogger(__name__)