+++ scifit/interfaces/generic.py
@@ -1,6 +1,5 @@
 import inspect
 import itertools
-import logging
 import numbers
 from collections.abc import Iterable
 
@@ -10,12 +9,12 @@ import numpy as np
 import pandas as pd
 from scipy import optimize, stats
 
+from scifit import logger
 from scifit.errors.base import *
+from scifit.interfaces.mixins import *
 
-logger = logging.getLogger(__name__)
 
-
-class FitSolverInterface:
+class FitSolverInterface(FitSolverMixin):
 
     """
     This class is an interface exposing clean way to fit a defined model to experimental data
@@ -65,109 +64,12 @@ class FitSolverInterface:
     Which defines a simple linear regression that can be fitted to experimental data.
     """
 
-    _data_keys = ["_xdata", "_ydata", "_sigma"]
-    _result_keys = ["_solution", "_minimize", "_yhat", "_loss", "_score", "_gof"]
+    _dimension = None
+    _data_keys = ("_xdata", "_ydata", "_sigma")
+    _result_keys = ("_solution", "_minimize", "_yhat", "_loss", "_score", "_gof", "_k2s")
 
-    def __init__(self, **kwargs):
-        """
-        Create a new instance of :class:`FitSolverInterface` and store parameters to pass them to :py:mod:`scipy`
-        afterwards when fitting experimental data.
-
-        :param kwargs: Dictionary of parameters to pass to :meth:`scipy.optimize.curve_fit`
-        """
-        self._configuration = kwargs
-
-    def configuration(self, **kwargs):
-        """
-        Return stored configuration updated by parameters passed to the method.
-
-        :param kwargs: extra parameters to update initially stored configuration
-        :return: Dictionary of parameters
-        """
-        return self._configuration | kwargs
-
-    def clean_results(self):
-        for key in self._result_keys:
-            self.__dict__.pop(key, None)
-
-    def store(self, xdata=None, ydata=None, sigma=None, data=None):
-        """
-        Validate and store features (variables), target and target uncertainties.
-
-        :param xdata: Features (variables) as a :code:`(n,m)` matrix
-        :param ydata: Target as a :code:`(n,)` matrix
-        :param sigma: Target uncertainties as :code:`(n,)` matrix
-        :param data: Full dataset including all xdata, ydata and sigma at once
-        :raise: Exception :class:`scifit.errors.base.InputDataError` if validation fails
-        """
-
-        # Partially import dataframe if provided, override with other fields
-        if data is not None:
-            if not isinstance(data, pd.DataFrame):
-                raise InputDataError(
-                    "Data must be of type DataFrame, got %s instead" % type(data)
-                )
-            if xdata is None:
-                xdata = data.filter(regex="^x[\d]+").values
-            if ydata is None:
-                ydata = data["y"].values
-            if sigma is None and "sy" in data.columns:
-                sigma = data["sy"].values
-
-        xdata = np.array(xdata)
-        ydata = np.array(ydata)
-
-        if xdata.ndim != 2:
-            raise InputDataError("Features must be a two dimensional array")
-
-        if xdata.shape[0] != ydata.shape[0]:
-            raise InputDataError(
-                "Incompatible shapes between x %s and y %s" % (xdata.shape, ydata.shape)
-            )
-
-        if not (
-            np.issubdtype(xdata.dtype, np.number)
-            and np.issubdtype(ydata.dtype, np.number)
-        ):
-            raise InputDataError("Input values must be numeric")
-
-        if np.any(np.isnan(xdata)) or np.any(np.isnan(ydata)):
-            raise InputDataError("Input values cannot contain missing data (NaN)")
-
-        if isinstance(sigma, Iterable):
-            sigma = np.array(sigma)
-            if sigma.shape != ydata.shape:
-                raise InputDataError(
-                    "Sigma as array must have the same shape as ydata %s, got %s instead"
-                    % (ydata.shape, sigma.shape)
-                )
-            if not np.issubdtype(sigma.dtype, np.number):
-                raise InputDataError(
-                    "All sigma must be numbers, got '%s' instead: %s"
-                    % (sigma.dtype, sigma)
-                )
-            if not np.all(np.isfinite(sigma)):
-                raise InputDataError("All sigma must be finite numbers: %s" % sigma)
-            if not np.all(sigma > 0.0):
-                raise InputDataError(
-                    "All sigma must be strictly positive numbers: %s" % sigma
-                )
-        elif isinstance(sigma, numbers.Number):
-            if sigma <= 0.0:
-                raise InputDataError("Sigma must be strictly positive")
-        elif sigma is None:
-            pass
-        else:
-            raise InputDataError("Sigma must be a number or an array of number")
-
-        self._xdata = xdata
-        self._ydata = ydata
-        self._sigma = sigma
-
-        logger.info(
-            "Stored data into solver: X%s, y%s, s=%s"
-            % (xdata.shape, ydata.shape, sigma is not None)
-        )
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
     @staticmethod
     def model(x, *parameters):
@@ -201,7 +103,7 @@ class FitSolverInterface:
         """
         Number of features (variables) inferred from experimental data shape
         """
-        return self._xdata.shape[1]
+        return self._dimension or self._xdata.shape[1]
 
     @property
     def m(self):
@@ -1918,3 +1820,15 @@ class FitSolverInterface:
                 axes.append(axe)
 
         return axes
+
+
+class FitSolver1D(FitSolverInterface):
+    _dimension = 1
+
+
+class FitSolver2D(FitSolverInterface):
+    _dimension = 2
+
+
+class FitSolver3D(FitSolverInterface):
+    _dimension = 3